Call Submission#deliver with pessimistic SQL locking
This is important because without the SQL locking if the same worker job
runs in parallel workers, multiple times, then any checks to see the
submission state could be incorrect due to updates which happing in the
other workers.

Comments in the spec can help explain what is going on here.