Implement Postgres search matching
Gives us word stemming and search ranking. Works by pulling out a list
of all keywords from the curated links. The keywords are matched with
the request to determine important words in the request. This word list
can then be used to generate a relevancy rank against the curated links
of which we need a greater than 0.75 match and we pick the top 3.