Rework how we queue the deliver submission worker
We are hitting an issue where the Sidekiq job runs too quickly and the
submission isn't retrievable as the DB write transaction hasn't
completed.

By moving the initial write outside the transaction we are ensuring it
completes before the next transaction is opened. This means the job will
always be able to find the queued submission and not exit early.