Add PublishedRequest
This is a model to store imported data from the disclosure log. They'll
be used in `FoiSuggestion` to provide automated suggestions for
previously asked questions.

We'll want to store the whole payload from the API response for the
request attributes, and then cache specific ones in their own column
later.

See http://guides.rubyonrails.org/active_record_postgresql.html for
notes on JSONB and Rails.

The block + Hash syntax in the factory for the payload attribute comes
from https://stackoverflow.com/a/32467854.