Clean PublishedRequest#summary when caching
We use the `summary` attribute when looking for suggestions to present
to the user, so cleaning it should help with search ranking.

Also means less processing at view render time, as now the helper is
just truncating, so a small performance boost too.

Fixes https://github.com/mysociety/foi-for-councils/issues/157