Update spec to use `update_all`
This needed in order to update to Rails 5.2.2 as
ActiveRecord::Persistence#update now [1] only handles single objects,
whereas it used to have a default `id = :all` [2]

Without this we get:
  Failure/Error: Submission.update(state: Submission::DELIVERED)
  ArgumentError: wrong number of arguments (given 1, expected 2)

Rubocop has to be disabled for this line as #update_all bypasses model
validation which could have unforeseen consequences. In this case its
fine as its in a single spec.

[1] https://api.rubyonrails.org/v5.2.2/classes/ActiveRecord/Persistence/ClassMethods.html#method-i-update
[2] https://api.rubyonrails.org/v5.2.1.1/classes/ActiveRecord/Persistence/ClassMethods.html#method-i-update